Add OATCommonListeners to handle player join events and update checking
